added Manuscripts

+++ bin/commands.rb
@@ -26,7 +26,7 @@ module Codex extend self
     # keep an array of errors
     errors = []
 
-    if $loaded_apps.instance_of? Array
+    if not $loaded_apps.empty?
 
       $loaded_apps.each do |app|
 
@@ -108,7 +108,7 @@ module Codex extend self
     errors = []
     restored_apps = []
 
-    if $loaded_apps.instance_of? Array
+    if not $loaded_apps.empty?
 
       $loaded_apps.each do |app|
 
@@ -175,37 +175,39 @@ module Codex extend self
   # This function will remove the symlink and COPY the codex file/folder to the local system
   def unlink
 
-    $loaded_apps.each do |app|
+    if not $loaded_apps.empty?
 
-      app['paths'].each do |path|
+      $loaded_apps.each do |app|
 
-        path = Codex.tildeToHomeFolder path
-        codex_path = Codex.getCodexPath(path)
-        type = Codex.getType(path)
+        app['paths'].each do |path|
 
-        begin
+          path = Codex.tildeToHomeFolder path
+          codex_path = Codex.getCodexPath(path)
+          type = Codex.getType(path)
 
-          # remove linked file on local system
-          if type === 'folder'
-            FileUtils.rm(path)
-            puts "Removing folder #{path}"
-            FileUtils.cp_r(codex_path, path)
-          elsif type === 'file'
-            FileUtils.rm(path)
-            # copy the codex file to the local system
-            FileUtils.copy(codex_path, path)
-            puts "Removing file #{path}"
+          begin
+
+            # remove linked file on local system
+            if type === 'folder'
+              FileUtils.rm(path)
+              # copy the codex file to the local system
+              FileUtils.cp_r(codex_path, path)
+            elsif type === 'file'
+              FileUtils.rm(path)
+              # copy the codex file to the local system
+              FileUtils.copy(codex_path, path)
+            end
+
+          rescue
+            # TODO: write rescue code to ensure no data was lost!
+            puts "\u2757 something went wrong! #{$!}"
           end
 
-        rescue
-          # TODO: write rescue code to ensure no data was lost!
-          puts "\u2757 something went wrong! #{$!}"
         end
 
       end
 
     end
-
   end
 
 end
